Week 3 - Added Start Menu
The game now has a start menu and a play button on that menu. The play
button on the game scene is gone and the code has been adjusted so that
the game automatically plays after the user presses the play from the
start menu or presses the restart button on the game over menu.