Week 4 - Added in Instructions
Question mark button was added and, if tapped on, shows, in a different
scene, the instructions on how to play. If the player goes back by
clicking on the “x” button, it will guide them back to their previous
scene (for now only start menu works but going back to game scene does
not).