Week 4 - Instructions & Pause Implemented
Now the user can click on the question mark to see instructions on how
to play the game in the start menu and in the game scene. Also, the
pause button works and, when pressed, a pause menu will appear with
three buttons (home, refresh, and play). All buttons work.